fix(server): break manage_operations circular init (TDZ unit fail)
Root cause of red CI unit on main: manage_operations imported
dispatch-chrome-action which re-imported manage_operations, so
BUILDER_APPROVAL_HANDLERS evaluated MANAGE_WATCHERS_ACTION_KEY while
manage_watchers was still in TDZ.

- Move waitForDeviceActionRun to device-action-wait.ts
- Lazy-init BUILDER_APPROVAL_HANDLERS
- Relax sdkSearch phrase-match test for live connector hits
